Prepare to move some assets around
