+++ package.json
@@ -24,6 +24,7 @@
 		"embla-carousel-react": "^8.3.0",
 		"framer-motion": "^11.11.9",
 		"gray-matter": "^4.0.3",
+		"image-size": "^1.1.1",
 		"lucide-react": "^0.453.0",
 		"next": "^15",
 		"next-themes": "^0.3.0",
