--- src/assets/json/gpg_key.json
@@ -1,3 +1,0 @@
-{
-	"pub_key": "-----BEGIN PGP PUBLIC KEY BLOCK-----mQINBGDCPEMBEADa81ZwBy8EzNvchHQZCJc7za4H8AA0Wsdknz54EUi7tbPwSuWjAqKS4QDrMeyLyzL97d9xVMYg5ezhSC/XtEsIy2h6Ral8+kV6a7K+z72juCj55Qm72LYEQXdx70gYGYK4LhUQa/3Ril2LdMiwEC3AL1d2pVFO7T5HN84JNWNeOHDhUeoHBPBHIwOFyJhTBqpWIcbNbnMQXK4s3dNIfD9oiQc204a3YeZIQ7D357H0NrANzypz/7aa0lGGz13W7GfHEMpVj4oemlN7btMxU7o39nPaSZVmOOMPbaJ2RgN2sx7u3QZwN32lJyVouo8xaVPb9f7USCchpc9qlZFbNw/yoNsqoptgjfyFwLR7IEUFZdaFRtfgl+fx49eHseKJlaKJNZaHo+pJOPLeZg5e8fP0VyIf/IgI3x7/dO+yLE14llnvPcWqZ8eh9bm+I/SRodq7w05CCSJSWcU0TP8ou3WBKPK3lKkXJIpo9xmE6HzI80OVGS6neRH98HqfC+8u0Hd0GvUizXejxTGsplMSge6wtQ4PHtqDlLpMcBnpBw7BO0hooMRFzKT6mE5PC9tUDdm2IeJCG9i8ycC8lRJWFTQinXO05D6GcWiqFeFYfhks1Hn3IfmkMJ6qo3/vli8kox3h00oEZSen0nPNsDKOamLmIIwsvoJdzZF6thvIkxDv6wARAQABtDdBbGV4IFdlc3Rlcm1hbiAoUGVyc29uYWwgRW1haWwpIDxhd2VzdGVybWFuMTZAbGl2ZS5jb20+iQJSBBMBCAA8FiEET2r+sIb4nimERZ/Cdu1dbQMBLjEFAmJFzXkCGwMFCwkIBwIDIgIBBhUKCQgLAgQWAgMBAh4HAheAAAoJEHbtXW0DAS4x3gYQAMnGSwr8AIkBQgaMZthC5OOgIfbsliv47nLAKzsUHzhJ+DLfNPWV1wVIzBATkqv3fmLIeeinHGd+zfCtS16a4rdLrwsRMYkokMMKXzbfsEBa/s1pP6QSuP7nAs9XfjQcbLQfEUf3hPIy8OcLOlWHs3W3icm+k30AAqBsammdmhJSBiNViTzplRz3om0nhk/VAOc4RZPlhmm2/QTaWprhIurzvJxp1iboLbLDB7Xx74Ae96eHTEcEa3W/7Dme5r/4hDdYIchL60rzS29JNAqeRR54doeuHh2UIHAUb/4RAazv3/o1MPjf+Q9MLXuXEB6MlTqCPME8DHPQGr6ekIiO/b0MEGh6DiEPHSEkG0KDzfBiBTlPvuTERDH9A8GbQ6V/5zKb/hwcEQYpdAL/6UhfrG/tUSkgkNzQtxFYepZmy6te9uNoa+VIAoAG/qEcpcv9E+VHtX1g4h00tP1EhoSj/pn+UlEn2HwCeG+6knT5i5XN+pPKXdlUGTR7ukgpFJkHEOyi0qZbZ93SPL4T/K746z1kCMFxHRMXi/9l9MIxwgp1QugVo9qL98rAbZgvoU6CiHQsnJ0H6XRD7kF6w7BHa0yFPm7zqnOXGjRcE3z0JFEe+RKkFDP+02Vs2HBHN+aldsvKIYtQdyBMGnqP0Fq7WDWkoC58xiIf+gS/SSd3vPg+tElBbGV4IFdlc3Rlcm1hbiAoV2luZG93cyBHbnVQRyBBbHRlcm5hdGl2ZSBLZXkpIDx3ZXN0ZXJhakByb3NlLWh1bG1hbi5lZHU+iQJOBBMBCAA4FiEET2r+sIb4nimERZ/Cdu1dbQMBLjEFAmDCPEMCGwMFCwkIBwIGFQoJCAsCBBYCAwECHgECF4AACgkQdu1dbQMBLjEh6A/+I5PCx1R1lzBMse0cks3BcCiMHw5siyT/atQ3vC0DKmm0aiIyVWis31Tmy5igIJIVDT+sH3j9BKxnP/EDXK5sFj45cBLltnuhMnWEXLFKrR01MPzuVMmL7K1mrlovNjkTRZ+ZLJVg5Lm9ZzRjYFWuF4unpqyHIquGSxKBxOQFPZhjOvYcIAK0NX67wXJfXGTGjAYrKdcBwIu6H4czNtLOQICgo5pOTsuOMyIBOGua2Ovvv++fFG8MFMSAPH+EUCI1wgL2cy7TSOfljOZl2IC/faWMYhsQ+6odh8K9Wsw1D7Pk7q2Kchgfayyf48ZgAPU7ofnmazTQ+JpS1kFnPWQJjF+ONJhjtCkdiK2fGyoSjvFFdnTg4e/vNBtiEr3Q+QNygQElBW/fdqBUr9Ia2P3UJAFPFE/wMKCAUi78M2SLYzzXb4O+eXqkBy8kTKywCv+9xF3BUipP9nLsDHWTzxGLBRKzqHb8zf0Fu1q4lvZpvjJ1fqBYZRFizdACDRHTPPWoBdRYIXea/Hbji0y8PZq0ch3mxSw+DDLV8OnC1k+66al36Fl9KvhVvYT3sJH7XW8SRSBOSp3rgo6m1VTBAnGnzSxXtvXBNu8H2+bTLJ83GNMWmsocxgrL46cMecp+wyTS/9PXnsLYUrypDYcVBjLTE5vXuGCeEbsgDG9DytkpMzGJAlEEEwEIADsCGwMFCwkIBwIGFQoJCAsCBBYCAwECHgECF4AWIQRPav6whvieKYRFn8J27V1tAwEuMQUCYkXYUwIZAQAKCRB27V1tAwEuMdDvD/9e8qIVaa2LtBL8dDdigLqCS+0JM3j7adeBrbLOEJS7fZSnivTUjMlMz6hNTpoYj1amAFAC1tQUfKrRElVWHzDAT47PNC9jmCcUyvJwXbW+ynOv8q9DIrDJkKlXCMXNhTSQePfIosxJwHq5+SPs0QnNnnxktz8Wwu861yOi+XnOCwi1F2hdbwQ1JVv8KdoUvGGMz9qFvyiUxdUAOJnC5xv/7oeSyt9tVflZ4h6wfzncRHIsUrG+yzDrWrJMbE6ixQqV7exQ5K50i5XoyFdM83L+Kt/0hlNy5K4MiZT6S2b6pPyhoS7UV7MviRURom2wjJ7twzmvJf0pekp/X3UbnOsn4jClp0OXUwytNP06NOCc+MwbV6KZNxwniauSM4x8nRr6g5RQHWINsULZbJjZAEJypNUCdOOiNvNCrkrbJfIpKjbitwLe2dD/dXrcAz+rdAU5tqm9kfAmGu1pkvNmK2Xa15epgI2FJ6OV0m4witEUiVG4w+K8Q5OzPq+OyeZEnWVm0oM6y3OPsLpm1gPWDs+3AuYFq4htVtsxMJQ45qqSN5ufP2TRB6AchQtU6ktLj0jeVdbysgJEQjXElqrqLmYd5ct5ImEaeuZQGUiaaYznhsf5oY7juT/+m5CNqhEfTEgCCQkX+hNSUWD1jH2OE9xvmk3EzQ+54QzAU295r6WFybQkQWxleCBXZXN0ZXJtYW4gPGZhY2VmdHdAb3V0bG9vay5jb20+iQJSBBMBCAA8FiEET2r+sIb4nimERZ/Cdu1dbQMBLjEFAmJFzlQCGwMFCwkIBwIDIgIBBhUKCQgLAgQWAgMBAh4HAheAAAoJEHbtXW0DAS4xsxEQAMA2sbwElkOacrnaAaYi+v6wreXFD9pawPAbhiqGLV5u0UqBRdQ2AaIUDooiy5/3lhc5nHpKQMsW3+vstIuo75hXhCCEjio4aKnJRF9EyIRfNiFp+acIQAB9tY7usIr9S0d7vOuEWDrqMUzk7PyTayLZ10kXba20ubJNumWliCAL4uzWMuxGJMXrP3qkwgKEuuUiXYnEm/9za03GVYUBTvIhKiIBk3Sz3B1kzHDHFb0JEkR3nYuAuFdMhfVgYNwi549DLq0GS48SiDaN9fG2dxgH8Xy3xAeUGg6XJ5VDCKb8lSbtEueWMZo13FC1SMiBCoVwQrYOkvUgDDt32nGZ+Kq0oYXfGFyjtaeKEUxR5Y479WCoDFAokw/nb4RuhffxCw/f0DPWly99pelRrX1wM3IHsYmMvbSduSCyBlGypX05XcF2OIfYPBwesVwg27wVznrUkqXEPWfDbXo1SFQbBcKJJqW7IjWQ0mfIo/tqgMwNeCIoGe79yrpdlpl+Og7x1fStYQqhtdfwSsf+U9BbbyeYeGxxMjSFLrXmpAmwhsg07AnlxswLSn+0eDLgxVdhMPQzjw4NESPu4gJlUlOpRQAoEuupRlns4rA6VbJf2OM9IHM+FNtS5si/pjMyow40BaBDmBGvH3tpS5Z7ybPqMfviO72a/XM9Ln+pYFADxTowtCFBbGV4IFdlc3Rlcm1hbiA8YWxleEBmYWNlZnR3LmRldj6JAlEEEwEIADsWIQRPav6whvieKYRFn8J27V1tAwEuMQUCYwvmFgIbAwULCQgHAgIiAgYVCgkICwIEFgIDAQIeBwIXgAAKCRB27V1tAwEuMcqZEAC3F0P4Ja0leVFE1QSXjgrMEGATMToNvhc5x/oSDTpofClzIWfy/RFw8euc1dLsU1SnA2s10LajPW/ltuXFUqb96P3znl3aPMQXFT01K0zscw0SzdZjwQ837U7lgY9EDoUGfvTkZNjplbdMDGxMDB9zp64m+MxgyPKpb4eH2UTx3W4Y0FkxRO/uj6TGto5d8d77Oz5NgsPJ3S8lI3XeGYf1g3ZsI6t/utfgjkd+HvZ5wU938FTYk0c1GWamyj3blvVggSfSEIKnT7b2folv3lLwKoH+wdgulTet/BT+wMdKCpx5UtOl94zhSTSJ3iIWTplJjoRq4TgJkMHUkUKmhwJN9DHgKtnBpab8bsrzGkfRr3aaz7kIja/DIZ5A+OlcEXcAKv159ZLygVajia7vtaYh/iNOMA6RsgZpz4rD52/61bwgttHZq1WKRjwsiHH8Rx3is2GGfN/Kk2KeAvOa+sAiTGWYoO/ubi11nFkOajvkJ/KEx8hgHSPtt1zRoBetwlt2BIH104rN8kixaANrWPBZBwHbDMsy99AXt6IcgXAqRuCrvTlIAYiqMh+TovWlCJhsHF9eRJEWK/SSMbZE9M02NXOX49kejzhHKzGo7pF0NukR+piXNkmuJuFkFlHmTpWFCnzjysmrQ+k+qNXWOxZ1/go37vJ02jN3m239vZFvbbkCDQRgwjxDARAAoCOj4ELLU+MY1Tld3qamwtGZhpPSamPXDvqURBJrNX883M+myPDWxi6th/Z2GJUrZk74nxoyB1a4pfWbTATKi63V7wKkOHPTTRuI9V+i4muctNSKZOqhm4kktx7wZDM8JVhyEj23+vwF6PSyY9LIOtjfOHzS9QNuKflS6fmIT9S0twaSdDSt3/PAFuz3ajc5GeklSYgXHm7MPfUbiS4rfELdGoz1sTQ91aXsab11i9cCbTYx1cM3H6a4uv49EeOP7Stx3NagQIVicyCcaYgbAfU3PxAbzRCdMQ5x8+Q5XvFvbWuTcSLLkaTnOmZijHTa8XXq+RrPGCnURiUECYP678Newz11WIb7vs9PnTGRMpZnMy/8Du2txtIsEmLQHfLzYFAvgS1vOPrdrjwSoypv5vjLDnfT5ACKLTis0naYa5nX/O8AYtYNuG9DUZCd/TQu7h4v6h2RKBTXjGPUoJkhznhM0DBwcNk5ZO+H0GoZInQmGAC10eA/eUz0Iy7UPKKB/rBh01/KyVqADgoiOumM8Znnd2FKS/n4JvNVc/nS0YOu9dKGzasK4j105Dz6UrsUZK9V3auja18BQVSUdVokhZ+slWpaz88U3e8HxzJ0Er0pnTx+g561pEXx9HQzNQey6E9dJ4eu0NlAjDnoaW8APAOpi5jHuef6Nia2xjQcjoUAEQEAAYkCNgQYAQgAIBYhBE9q/rCG+J4phEWfwnbtXW0DAS4xBQJgwjxDAhsMAAoJEHbtXW0DAS4xvX4P/2vNQW1as2NRwVaDuS9mFeIEFNarQt+vhy4jwwaSaZXM2ubw6iueQXnS6UtKS7C07y5tuyqnVPT/Im1zcWBWyLoUIOHwnQ/PbCN5WNRWX68vAfFtzJb4mzZRiNqhgAT1sfVpRH+UIlCM7QIzc+T4wINrjY2DXC8BhMzmBODVpqjdw7QBUZ2wVQguLLmo77v/tOLp3BW5NBYNvU+VBvrCJZdtIBxNe+oafsiawxjmPamunrtZa85O0kBacAF0XfdJFgSovNg9TfAKnLJOx9z+FcazWxvbRQEnu/KDILKKf28jsA8KkiJEQiyt/BIckDQeq+XgpcM2mMZsLnCiLMkW4SoDcxT16D+BliDTM9IGhKDW3Vr6lvVAJgHpy0e9oUKM4TGSbsnuZMmzlI5LA7H3zkt7EOPczlGDNr0KmAQvgebOvB5ReU90ubz6Yb39aStdUVA0VSFTP0DT2pK8AeAwufD2EWmKAzUQIvCeWjFIiXurNYGH7n43ckIKWkvD7XQ+SkJDW9JJSTvnHJctFEUXb4HQ73nPanCgV9zU7VIIPyjlqkWfGnX/l5MewyiNu0XT8x/FQUwhcbRRP02BSYjDkWUwHZAPQ8L/C33V6bIzusuIuU5o2BEvkdsbJqtGBs2j4iDYmGeT61arZsZdtbqVPh6CDOIkiIdFdCxx3euvR3SQ=aLQx-----END PGP PUBLIC KEY BLOCK-----"
-}
