--- src/assets/misc/Alex Westerman_0x03012E31_public.asc
@@ -1,52 +1,0 @@
------BEGIN PGP PUBLIC KEY BLOCK-----
-
-mQINBGDCPEMBEADa81ZwBy8EzNvchHQZCJc7za4H8AA0Wsdknz54EUi7tbPwSuWj
-AqKS4QDrMeyLyzL97d9xVMYg5ezhSC/XtEsIy2h6Ral8+kV6a7K+z72juCj55Qm7
-2LYEQXdx70gYGYK4LhUQa/3Ril2LdMiwEC3AL1d2pVFO7T5HN84JNWNeOHDhUeoH
-BPBHIwOFyJhTBqpWIcbNbnMQXK4s3dNIfD9oiQc204a3YeZIQ7D357H0NrANzypz
-/7aa0lGGz13W7GfHEMpVj4oemlN7btMxU7o39nPaSZVmOOMPbaJ2RgN2sx7u3QZw
-N32lJyVouo8xaVPb9f7USCchpc9qlZFbNw/yoNsqoptgjfyFwLR7IEUFZdaFRtfg
-l+fx49eHseKJlaKJNZaHo+pJOPLeZg5e8fP0VyIf/IgI3x7/dO+yLE14llnvPcWq
-Z8eh9bm+I/SRodq7w05CCSJSWcU0TP8ou3WBKPK3lKkXJIpo9xmE6HzI80OVGS6n
-eRH98HqfC+8u0Hd0GvUizXejxTGsplMSge6wtQ4PHtqDlLpMcBnpBw7BO0hooMRF
-zKT6mE5PC9tUDdm2IeJCG9i8ycC8lRJWFTQinXO05D6GcWiqFeFYfhks1Hn3Ifmk
-MJ6qo3/vli8kox3h00oEZSen0nPNsDKOamLmIIwsvoJdzZF6thvIkxDv6wARAQAB
-tElBbGV4IFdlc3Rlcm1hbiAoV2luZG93cyBHbnVQRyBBbHRlcm5hdGl2ZSBLZXkp
-IDx3ZXN0ZXJhakByb3NlLWh1bG1hbi5lZHU+iQJOBBMBCAA4FiEET2r+sIb4nimE
-RZ/Cdu1dbQMBLjEFAmDCPEMCGwMFCwkIBwIGFQoJCAsCBBYCAwECHgECF4AACgkQ
-du1dbQMBLjEh6A/+I5PCx1R1lzBMse0cks3BcCiMHw5siyT/atQ3vC0DKmm0aiIy
-VWis31Tmy5igIJIVDT+sH3j9BKxnP/EDXK5sFj45cBLltnuhMnWEXLFKrR01MPzu
-VMmL7K1mrlovNjkTRZ+ZLJVg5Lm9ZzRjYFWuF4unpqyHIquGSxKBxOQFPZhjOvYc
-IAK0NX67wXJfXGTGjAYrKdcBwIu6H4czNtLOQICgo5pOTsuOMyIBOGua2Ovvv++f
-FG8MFMSAPH+EUCI1wgL2cy7TSOfljOZl2IC/faWMYhsQ+6odh8K9Wsw1D7Pk7q2K
-chgfayyf48ZgAPU7ofnmazTQ+JpS1kFnPWQJjF+ONJhjtCkdiK2fGyoSjvFFdnTg
-4e/vNBtiEr3Q+QNygQElBW/fdqBUr9Ia2P3UJAFPFE/wMKCAUi78M2SLYzzXb4O+
-eXqkBy8kTKywCv+9xF3BUipP9nLsDHWTzxGLBRKzqHb8zf0Fu1q4lvZpvjJ1fqBY
-ZRFizdACDRHTPPWoBdRYIXea/Hbji0y8PZq0ch3mxSw+DDLV8OnC1k+66al36Fl9
-KvhVvYT3sJH7XW8SRSBOSp3rgo6m1VTBAnGnzSxXtvXBNu8H2+bTLJ83GNMWmsoc
-xgrL46cMecp+wyTS/9PXnsLYUrypDYcVBjLTE5vXuGCeEbsgDG9DytkpMzG5Ag0E
-YMI8QwEQAKAjo+BCy1PjGNU5Xd6mpsLRmYaT0mpj1w76lEQSazV/PNzPpsjw1sYu
-rYf2dhiVK2ZO+J8aMgdWuKX1m0wEyout1e8CpDhz000biPVfouJrnLTUimTqoZuJ
-JLce8GQzPCVYchI9t/r8Bej0smPSyDrY3zh80vUDbin5Uun5iE/UtLcGknQ0rd/z
-wBbs92o3ORnpJUmIFx5uzD31G4kuK3xC3RqM9bE0PdWl7Gm9dYvXAm02MdXDNx+m
-uLr+PRHjj+0rcdzWoECFYnMgnGmIGwH1Nz8QG80QnTEOcfPkOV7xb21rk3Eiy5Gk
-5zpmYox02vF16vkazxgp1EYlBAmD+u/DXsM9dViG+77PT50xkTKWZzMv/A7trcbS
-LBJi0B3y82BQL4Etbzj63a48EqMqb+b4yw530+QAii04rNJ2mGuZ1/zvAGLWDbhv
-Q1GQnf00Lu4eL+odkSgU14xj1KCZIc54TNAwcHDZOWTvh9BqGSJ0JhgAtdHgP3lM
-9CMu1Dyigf6wYdNfyslagA4KIjrpjPGZ53dhSkv5+CbzVXP50tGDrvXShs2rCuI9
-dOQ8+lK7FGSvVd2ro2tfAUFUlHVaJIWfrJVqWs/PFN3vB8cydBK9KZ08foOetaRF
-8fR0MzUHsuhPXSeHrtDZQIw56GlvADwDqYuYx7nn+jYmtsY0HI6FABEBAAGJAjYE
-GAEIACAWIQRPav6whvieKYRFn8J27V1tAwEuMQUCYMI8QwIbDAAKCRB27V1tAwEu
-Mb1+D/9rzUFtWrNjUcFWg7kvZhXiBBTWq0Lfr4cuI8MGkmmVzNrm8OornkF50ulL
-SkuwtO8ubbsqp1T0/yJtc3FgVsi6FCDh8J0Pz2wjeVjUVl+vLwHxbcyW+Js2UYja
-oYAE9bH1aUR/lCJQjO0CM3Pk+MCDa42Ng1wvAYTM5gTg1aao3cO0AVGdsFUILiy5
-qO+7/7Ti6dwVuTQWDb1PlQb6wiWXbSAcTXvqGn7ImsMY5j2prp67WWvOTtJAWnAB
-dF33SRYEqLzYPU3wCpyyTsfc/hXGs1sb20UBJ7vygyCyin9vI7APCpIiREIsrfwS
-HJA0Hqvl4KXDNpjGbC5woizJFuEqA3MU9eg/gZYg0zPSBoSg1t1a+pb1QCYB6ctH
-vaFCjOExkm7J7mTJs5SOSwOx985LexDj3M5Rgza9CpgEL4HmzrweUXlPdLm8+mG9
-/WkrXVFQNFUhUz9A09qSvAHgMLnw9hFpigM1ECLwnloxSIl7qzWBh+5+N3JCClpL
-w+10PkpCQ1vSSUk75xyXLRRFF2+B0O95z2pwoFfc1O1SCD8o5apFnxp1/5eTHsMo
-jbtF0/MfxUFMIXG0UT9NgUmIw5FlMB2QD0PC/wt91emyM7rLiLlOaNgRL5HbGyar
-RgbNo+Ig2Jhnk+tWq2bGXbW6lT4eggziJIiHRXQscd3rr0d0kA==
-=7Vwt
------END PGP PUBLIC KEY BLOCK-----
